convert other views and remove old extension view
