--- src/view/components/extensionViewContext.jsx
@@ -1,3 +1,0 @@
-import React from "react";
-
-export default React.createContext();
